refactor: Move storage to route group, remove searchParams prop
